Fix compiler warnings and do some formatting along the way

+++ ovc5/software/camera_device_driver/src/i2c_driver.cpp
@@ -53,7 +53,7 @@ int I2CDriver::writeRegister(const regop_t& regop)
   assert(initialized());
   std::vector<uint8_t> payload(reg_size + 1); // We need to add one byte for address
   payload[0] = regop.addr & 0xFF;
-  for (size_t i = 1; i <= reg_size; ++i)
+  for (int i = 1; i <= reg_size; ++i)
     payload[i] = regop.i32 >> (8 * (reg_size - i));
   return i2c_smbus_write_i2c_block_data(i2c_fd, regop.addr >> 8, payload.size(), &payload[0]); 
 }
